Refactor backoffice repair presentation
